Move QueryColumnPicker into querying/common/components
