+++ frontend/src/metabase/querying/common/components/QueryColumnPicker/ColumnBucketPickerPopover/BinningStrategyPickerPopover.tsx
