+++ frontend/src/metabase/querying/common/components/QueryColumnPicker/QueryColumnPicker.module.css
