+++ frontend/src/metabase/querying/components/expressions/CombineColumns/ColumnAndSeparatorRow/ColumnInput.tsx
@@ -3,7 +3,7 @@ import type { FocusEvent, KeyboardEvent, MouseEvent } from "react";
 import { useMemo, useRef, useState } from "react";
 import { t } from "ttag";
 
-import { QueryColumnPicker } from "metabase/common/components/QueryColumnPicker";
+import { QueryColumnPicker } from "metabase/querying/common/components/QueryColumnPicker";
 import { Button, FocusTrap, Icon, Input, Popover } from "metabase/ui";
 import { color } from "metabase/ui/colors";
 import { getThemeOverrides } from "metabase/ui/theme";
