+++ frontend/src/metabase/querying/notebook/components/BreakoutStep/BreakoutStep.tsx
@@ -1,10 +1,10 @@
 import { useMemo } from "react";
 import { t } from "ttag";
 
-import { QueryColumnPicker } from "metabase/common/components/QueryColumnPicker";
 import { useLocale } from "metabase/common/hooks";
 import { useTranslateContent } from "metabase/content-translation/hooks";
 import { PLUGIN_CONTENT_TRANSLATION } from "metabase/plugins";
+import { QueryColumnPicker } from "metabase/querying/common/components/QueryColumnPicker";
 import * as Lib from "metabase-lib";
 
 import type { NotebookStepProps } from "../../types";
